Comment amti and delete AME
We want to make this repo able to be installed and build from robotology-superbuild
It seeems that installing amti driver had issues in the past and needs to be tested again so for temporary preventive measure we are disabling from cmake list.
On the other hand AME never got any support from manufacturers and did not work, we currently do not have a sensor anymore so we should delete it.